Time: 0 ms (100.00%) | Memory: 8.3 MB (6.44%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -48,7 +48,7 @@ class Solution {
 public:
     double myPow(double x, int n) {
 		double s = 1;
-		s = (double)pow(x, n);
+		s = pow(x, n);
         return s ;
     }
 };
